tvads_rag/requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
openai>=1.0.0
httpx>=0.27.0
google-genai>=0.1.0
tqdm>=4.66.1
pydantic>=2.5.0
pytest>=7.4.0
cohere>=5.5.0
streamlit>=1.30.0
supabase>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
